requirements.txt
streamlit==1.37.0 
openai==1.42.0 
python-dotenv==1.0.1 
